pyproject.toml
[project]
name = "rf_processtree"
dynamic = ["version"]

description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = ["robotframework"]

[dependency-groups]
dev = ["ipython", "xonsh[full]", "ruff"]

[build-system]
requires = ["hatchling", "versioningit"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "versioningit"